@@ -111,10 +111,13 @@
     "webpack": "5.74.0"
   },
   "resolutions": {
-    "@types/react": "^16.9.23",
-    "@types/react-dom": "^16.9.5",
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0",
+    "@types/react": "~18.2.0",
+    "@types/react-dom": "~18.2.0",
+    "react": "~18.2.0",
+    "react-dom": "~18.2.0",
+    "visyn_core": "git+ssh://git@github.com/datavisyn/visyn_core#react18",
+    "visyn_scripts": "git+ssh://git@github.com/datavisyn/visyn_scripts#react18",
+    "typescript": "~5.0.4",
     "webpack": "5.74.0"
   },
   "visyn": {
