@@ -1,12 +1,10 @@
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { VisynAppProvider } from 'visyn_core';
 import { MainApp } from './MainApp';
 
-// create a new instance of the app
-ReactDOM.render(
+createRoot(document.getElementById('main')).render(
   <VisynAppProvider appName="Demo App">
     <MainApp />
   </VisynAppProvider>,
-  document.getElementById('main'),
 );
