@@ -71,7 +71,7 @@ export class IDTypeUtils {
 
   static async isIDType(name: string, index: number, data: any[], accessor: (row: any) => string, sampleSize: number) {
     // first check if it is number then it cant be an IDType
-    const isNumber = PHOVEA_IMPORTER_ValueTypeUtils.numerical().isType(name, index, data, accessor, sampleSize);
+    const isNumber = await PHOVEA_IMPORTER_ValueTypeUtils.numerical().isType(name, index, data, accessor, sampleSize);
     if (isNumber > 0.8) {
       // pretty sure it is a number
       return 0;
